requirements.txt
langchain
langchain-community
CTransformers
sentence_transformers
fastapi
jinja2
aiofiles
PyPDF2
python-multipart
faiss-cpu
huggingface_hub